@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <ApplicationIcon />
     <OutputType>Exe</OutputType>
     <StartupObject />
@@ -12,20 +12,7 @@
     <Content Remove="Data\**" />
     <EmbeddedResource Remove="Data\**" />
     <None Remove="Data\**" />
-  </ItemGroup>
-       
-  <ItemGroup>
-    <PackageReference Include="Autofac" Version="4.8.1" />
-    <PackageReference Include="Autofac.Extensions.DependencyInjection" Version="4.3.0" />
-    <PackageReference Include="LazyCache.AspNetCore" Version="2.0.0-beta03" />
-    <PackageReference Include="Microsoft.AspNetCore.All" Version="2.1.3" />
-    <PackageReference Include="Quartz" Version="3.0.6" />
-    <PackageReference Include="Serilog.Extensions.Logging" Version="2.0.2" />
-    <PackageReference Include="Serilog.Settings.Configuration" Version="2.6.1" />
-    <PackageReference Include="Serilog.Sinks.Async" Version="1.3.0" />
-    <PackageReference Include="Serilog.Sinks.RollingFile" Version="3.3.0" />
-    <PackageReference Include="ZKWeb.System.Drawing" Version="4.0.1" />  
-  </ItemGroup>
+  </ItemGroup>     
 
   <ItemGroup>
     <ProjectReference Include="..\AspNetCore.Identity.LiteDB\AspNetCore.Identity.LiteDB.csproj" />
@@ -33,7 +20,6 @@
     <ProjectReference Include="..\MachinaTrader.Exchanges\MachinaTrader.Exchanges.csproj" />
     <ProjectReference Include="..\MachinaTrader.Globals\MachinaTrader.Globals.csproj" />
     <ProjectReference Include="..\MachinaTrader.Notifications\MachinaTrader.Notifications.csproj" />
-    <ProjectReference Include="..\MachinaTrader.Plugin.Sample\MachinaTrader.Plugin.Sample.csproj" />
     <ProjectReference Include="..\MachinaTrader.Strategies\MachinaTrader.Strategies.csproj" />
     <ProjectReference Include="..\MachinaTrader.Data.LiteDB\MachinaTrader.Data.LiteDB.csproj" />
     <ProjectReference Include="..\MachinaTrader.Data.MongoDB\MachinaTrader.Data.MongoDB.csproj" />
@@ -42,5 +28,16 @@
 
   <ItemGroup>
     <Folder Include="SimulationExchanges\" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Autofac.Extensions.DependencyInjection" Version="7.2.0" />
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="3.1.22" />
+    <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="3.1.22" />
+    <PackageReference Include="Microsoft.AspNetCore.Mvc.Razor.RuntimeCompilation" Version="3.1.22" />
+    <PackageReference Include="MongoDB.Driver" Version="2.14.1" />
+    <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
+    <PackageReference Include="Quartz" Version="3.3.3" />
+    <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="6.15.1" />
   </ItemGroup>
 </Project>
