@@ -146,9 +146,7 @@ namespace MachinaTrader.Globals
             {
                 //Init Global Config with default currency array
                 Global.Configuration = MergeObjects.MergeCsDictionaryAndSave(new MainConfig(), Global.DataPath + "/MainConfig.json").ToObject<MainConfig>();
-                Global.Configuration.TradeOptions.MarketBlackList = new List<string> { };
-                Global.Configuration.TradeOptions.OnlyTradeList = new List<string> { "ETHUSD" };
-                Global.Configuration.TradeOptions.AlwaysTradeList = new List<string> { "ETHUSD" };
+                Global.Configuration.TradeOptions.TradeAssets = "ETH,USD";
                 var defaultExchangeOptions = new ExchangeOptions
                 {
                     Exchange = Exchange.Binance,
